pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Sean Ross <sean@rossdev.io>"]
description = ""
name = "Portfolio"
version = "0.1.0"
[tool.poetry.dependencies]
Brotli = "^1.0.9"
Django = "^3.2.12"
Markdown = "^3.3.6"
dj-database-url = "^0.5.0"
django-markdownx = "^4.0.0b1"
gunicorn = "^20.1.0"
python = "^3.8"
sqlparse = "^0.4.2"
whitenoise = "^5.3.0"
django-secured-fields = "^0.3.1"
[tool.poetry.dev-dependencies]
